Adding a small program to parser the Javadocs for parameter names
The parser in the Xamarin.Android binding generator is not handling this new version of Facebook SDKs.

The plan is to run this program against all the sdks that are causing a problem and commit the transform files.

This can also be added as a pre-step to the libs target after the externals target is run.